break up package access test file into three: global, scoped public, scoped private
